% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_records.R
\name{download_records}
\alias{download_records}
\title{Donwload records from an OpenDataSoft portal}
\usage{
download_records(
  server,
  api = "/api/records/1.0/download/",
  proxy = TRUE,
  ...,
  verbosity = NULL,
  wrapper = data.table::as.data.table
)
}
\arguments{
\item{server}{The URL of the OpenDataSoft portal.}

\item{api}{Path for the API to use.}

\item{proxy}{Use or not a proxy server. If \code{TRUE},
an environment variable \code{PROXY_URL} must be set (along with \code{PROXY_PORT}, \code{PROXY_ID}, and \code{PROXY_PWD}).
Altrenatively a \code{list} can be used with names corresponding to arguments of \code{\link[httr2:req_proxy]{httr2::req_proxy()}}.}

\item{...}{Arguments passed to \code{\link[httr2:req_url]{httr2::req_url_query()}}.}

\item{verbosity}{How much information to print? This is a wrapper
around \code{\link[httr2:req_verbose]{req_verbose()}} that uses an integer to control verbosity:
\itemize{
\item \code{0}: no output
\item \code{1}: show headers
\item \code{2}: show headers and bodies
\item \code{3}: show headers, bodies, and curl status messages.
}

Use \code{\link[httr2:with_verbosity]{with_verbosity()}} to control the verbosity of requests that
you can't affect directly.}

\item{wrapper}{Function to wrap output data into, default is to return a \code{data.table}.}
}
\value{
Result of the request.
}
\description{
Donwload records from an OpenDataSoft portal
}
